+++ src/main/java/laddergame/controller/LadderController.java
@@ -2,16 +2,16 @@ package laddergame.controller;
 
 import java.util.List;
 import java.util.function.Supplier;
-import laddergame.domain.BooleanGenerator;
-import laddergame.domain.Height;
-import laddergame.domain.Ladder;
-import laddergame.domain.LadderGame;
-import laddergame.domain.LadderLinesGenerator;
 import laddergame.domain.LadderResult;
-import laddergame.domain.Line;
-import laddergame.domain.NamesWithItem;
 import laddergame.domain.PersonalNames;
-import laddergame.domain.Width;
+import laddergame.domain.game.LadderGame;
+import laddergame.domain.game.NamesWithItem;
+import laddergame.domain.ladder.Height;
+import laddergame.domain.ladder.Ladder;
+import laddergame.domain.ladder.Width;
+import laddergame.domain.ladder.line.BooleanGenerator;
+import laddergame.domain.ladder.line.LadderLinesGenerator;
+import laddergame.domain.ladder.line.Line;
 import laddergame.view.InputView;
 import laddergame.view.LadderFormGenerator;
 import laddergame.view.OutputView;
