+++ src/main/java/laddergame/domain/ladder/Height.java
@@ -1,4 +1,4 @@
-package laddergame.domain;
+package laddergame.domain.ladder;
 
 public class Height {
     private final int value;
