+++ src/main/java/laddergame/domain/ladder/Ladder.java
@@ -1,13 +1,14 @@
-package laddergame.domain;
+package laddergame.domain.ladder;
 
 import java.util.Collections;
 import java.util.List;
+import laddergame.domain.ladder.line.Line;
 
 public class Ladder {
     private final Width width;
     private final Height height;
 
-    private List<Line> lines;
+    private final List<Line> lines;
 
     public Ladder(final Width width, final Height height, final List<Line> lines) {
         this.width = width;
