+++ src/main/java/laddergame/domain/ladder/Width.java
@@ -1,4 +1,4 @@
-package laddergame.domain;
+package laddergame.domain.ladder;
 
 public class Width {
     private final int value;
