+++ src/main/java/laddergame/domain/ladder/line/BooleanGenerator.java
@@ -1,4 +1,4 @@
-package laddergame.domain;
+package laddergame.domain.ladder.line;
 
 @FunctionalInterface
 public interface BooleanGenerator {
