+++ src/main/java/laddergame/domain/ladder/line/RandomBooleanGenerator.java
@@ -1,4 +1,4 @@
-package laddergame.domain;
+package laddergame.domain.ladder.line;
 
 import java.util.Random;
