+++ src/test/java/laddergame/domain/LadderTest.java
@@ -6,6 +6,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import java.util.List;
+import laddergame.domain.ladder.Height;
+import laddergame.domain.ladder.Ladder;
+import laddergame.domain.ladder.Width;
+import laddergame.domain.ladder.line.Line;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
