+++ src/test/java/laddergame/domain/WidthTest.java
@@ -3,6 +3,7 @@ package laddergame.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import laddergame.domain.ladder.Width;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
